display_init copies digit_pins buffer

+++ code/shared/7-segment.h
@@ -15,7 +15,7 @@ void display_init_single(display_t *display, GPIO_TypeDef *segment_port, uint16_
 // initialise 7-segment display structure with arbitrary (> 1) digits.
 // provide port for segment pins, segment pins a-g, decimal pin, digit port, number of digits, digit pins,
 // and if the display is common anode.
-void display_init(display_t *display, GPIO_TypeDef *segment_port, uint16_t segment_pins[SEGMENT_COUNT], uint16_t decimal_pin, GPIO_TypeDef *digit_port, uint8_t digit_count, uint8_t *digit_pins, bool is_anode);
+void display_init(display_t *display, GPIO_TypeDef *segment_port, uint16_t segment_pins[SEGMENT_COUNT], uint16_t decimal_pin, GPIO_TypeDef *digit_port, uint8_t digit_count, uint16_t *digit_pins, bool is_anode);
 					
 // display given number to display, provide decimal point bit pattern (0x1 = dp1, 0x2 = dp2, 0x4 = dp3, ...)
 void display_write(display_t *display, uint64_t value, uint32_t decimal_points);
